+++ src/Calendars/EventParticipations.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace Swe\SpaceSDK\Calendars;
+
+use GuzzleHttp\Exception\GuzzleException;
+use Swe\SpaceSDK\AbstractApi;
+use Swe\SpaceSDK\Exception\MissingArgumentException;
+
+/**
+ * Class EventParticipations
+ *
+ * @package Swe\SpaceSDK\Calendars
+ * @author Luca Braun <l.braun@s-w-e.com>
+ */
+class EventParticipations extends AbstractApi
+{
+    /**
+     * Update RSVP / calendar event participation status for a calendar event attached to an article.
+     *
+     * @param string $id
+     * @param array $data
+     * @param array $response
+     * @return array
+     * @throws GuzzleException
+     * @throws MissingArgumentException
+     */
+    public function updateEventParticipation(string $id, array $data, array $response = []): array
+    {
+        $uri = 'calendars/event-participations/{id}';
+        $required = [
+            'newStatus' => self::TYPE_STRING,
+        ];
+        $this->throwIfInvalid($required, $data);
+        $uriArguments = [
+            'id' => $id,
+        ];
+
+        return $this->client->patch($this->buildUrl($uri, $uriArguments), $data, $response);
+    }
+}