+++ src/Calendars/Events.php
@@ -1,0 +1,55 @@
+<?php
+
+namespace Swe\SpaceSDK\Calendars;
+
+use GuzzleHttp\Exception\GuzzleException;
+use Swe\SpaceSDK\AbstractApi;
+use Swe\SpaceSDK\Exception\MissingArgumentException;
+
+/**
+ * Class Events
+ *
+ * @package Swe\SpaceSDK\Calendars
+ * @author Luca Braun <l.braun@s-w-e.com>
+ */
+class Events extends AbstractApi
+{
+    /**
+     * Get calendar events attached to an article in a specific time period.
+     *
+     * @param array $request
+     * @param array $response
+     * @return array
+     * @throws GuzzleException
+     * @throws MissingArgumentException
+     */
+    public function getAllEvents(array $request, array $response = []): array
+    {
+        $uri = 'calendars/events';
+        $required = [
+            'dateFrom' => self::TYPE_DATE,
+            'dateTo' => self::TYPE_DATE,
+        ];
+        $this->throwIfInvalid($required, $request);
+
+        return $this->client->get($this->buildUrl($uri), $response, $request);
+    }
+
+    /**
+     * Get a calendar event attached to an article.
+     *
+     * @param string $id
+     * @param array $response
+     * @return array
+     * @throws GuzzleException
+     */
+    public function getEvent(string $id, array $response = []): array
+    {
+        $uri = 'calendars/events/{id}';
+        $uriArguments = [
+            'id' => $id,
+        ];
+
+        return $this->client->get($this->buildUrl($uri, $uriArguments), $response);
+    }
+}