+++ src/Calendars/Holidays.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Swe\SpaceSDK\Calendars;
+
+use GuzzleHttp\Exception\GuzzleException;
+use Swe\SpaceSDK\AbstractApi;
+use Swe\SpaceSDK\Exception\MissingArgumentException;
+
+/**
+ * Class Holidays
+ *
+ * @package Swe\SpaceSDK\Calendars
+ * @author Luca Braun <l.braun@s-w-e.com>
+ */
+class Holidays extends AbstractApi
+{
+    /**
+     * Get/search holidays. Parameters are applied as 'AND' filters.
+     *
+     * @param array $request
+     * @param array $response
+     * @return array
+     * @throws GuzzleException
+     * @throws MissingArgumentException
+     */
+    public function getAllHolidays(array $request, array $response = []): array
+    {
+        $uri = 'calendars/holidays';
+        $required = [
+            'dateFrom' => self::TYPE_DATE,
+            'dateTo' => self::TYPE_DATE,
+        ];
+        $this->throwIfInvalid($required, $request);
+
+        return $this->client->get($this->buildUrl($uri), $response, $request);
+    }
+}